requirements.txt
Django==5.2.4
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
plotly==5.24.1
dash==2.18.1
dash-bootstrap-components==1.6.0
pandas==2.2.2
numpy==1.26.4
pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0